Move shard classes into their own package
