+++ src/main/java/com/avairebot/orion/plugin/JavaPlugin.java
@@ -1,7 +1,7 @@
 package com.avairebot.orion.plugin;
 
 import com.avairebot.orion.Orion;
-import com.avairebot.orion.OrionShard;
+import com.avairebot.orion.shard.OrionShard;
 import com.avairebot.orion.cache.CacheManager;
 import com.avairebot.orion.commands.CategoryHandler;
 import com.avairebot.orion.commands.CommandHandler;
